// apps/desktop/ui/src/data/generated/CompactionCount.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { CompactionEvent } from "./CompactionEvent";
import type { CompactionReason } from "./CompactionReason";

export type CompactionCount = { "state": "count", count: number, events: Array<CompactionEvent>, } | { "state": "unknown", reason: CompactionReason, };